storage: move SourceData to its own module + impl From<Partitioned...>
Using From/TryFrom to interopate Partitioned and SourceData is
convenient/idiomatic. To make that conversion more well structured,
move SourceData into its own module, and define the conversion
there. This also provides a sane place for other SourceData
conversions in the future.
