+++ src/storage-client/src/types/sources/data.proto
@@ -1,0 +1,22 @@
+// Copyright Materialize, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+syntax = "proto3";
+
+import "repr/src/row.proto";
+import "storage-client/src/types/errors.proto";
+
+package mz_storage_client.types.sources.data;
+
+message ProtoSourceData {
+    oneof kind {
+        mz_repr.row.ProtoRow ok = 1;
+        errors.ProtoDataflowError err = 2;
+    }
+}
