+++ src/storage/src/source/reclock/compat.rs
@@ -30,7 +30,8 @@ use mz_persist_client::write::WriteHandle;
 use mz_persist_client::Upper;
 use mz_repr::{Datum, Diff, GlobalId, Row};
 use mz_storage_client::controller::CollectionMetadata;
-use mz_storage_client::types::sources::{MzOffset, SourceData};
+use mz_storage_client::types::sources::data::SourceData;
+use mz_storage_client::types::sources::MzOffset;
 use mz_storage_client::util::antichain::{MutableOffsetAntichain, OffsetAntichain};
 use mz_timely_util::order::Partitioned;
